UPD: Обработана ошибка SelectNote = null

+++ ToDoCal/ViewModels/SelectedDateNotesViewModel.cs
@@ -88,8 +88,8 @@ namespace ToDoCal.ViewModels
             {
                 return new DelegateCommand(() =>
                 {
-                   
-                    SelectedNotePage();
+                   if(SelectNote is not null)
+                        SelectedNotePage();
                 });
             }
         }
@@ -103,7 +103,7 @@ namespace ToDoCal.ViewModels
             {
                 return new DelegateCommand(() =>
                 {
-                    if (SelectNote.Is_Task)
+                    if (SelectNote is not null && SelectNote.Is_Task)
                     {
                         Note.Edit_Note(SelectNote, null, null, null, "В процесcе");
                         UpdateNotes();
@@ -119,7 +119,7 @@ namespace ToDoCal.ViewModels
             {
                 return new DelegateCommand(() =>
                 {
-                    if (SelectNote.Is_Task)
+                    if (SelectNote is not null && SelectNote.Is_Task)
                     {
                         Note.Edit_Note(SelectNote, null, null, null, "Выполнено");
                         UpdateNotes();
@@ -135,7 +135,7 @@ namespace ToDoCal.ViewModels
             {
                 return new DelegateCommand(() =>
                 {
-                    if (SelectNote.Is_Task)
+                    if (SelectNote is not null && SelectNote.Is_Task)
                     {
                         Note.Edit_Note(SelectNote, null, null, null, "Брошено");
                         UpdateNotes();
@@ -151,7 +151,7 @@ namespace ToDoCal.ViewModels
             {
                 return new DelegateCommand(() =>
                 {
-                    if (MessageBox.Show("Вы точно хотите удалить заметку", "Подтвержение", MessageBoxButton.YesNo, MessageBoxImage.Information) == MessageBoxResult.Yes)
+                    if (SelectNote is not null && MessageBox.Show("Вы точно хотите удалить заметку", "Подтвержение", MessageBoxButton.YesNo, MessageBoxImage.Information) == MessageBoxResult.Yes)
                     {
 
                         Note.Delete_Note(SelectNote);
